+++ changelog.md
@@ -12,6 +12,7 @@
 - Add `silent` parameter to `Client.interaction_response_message_create`.
 - Add `create_user_from_thread_user_data` to help with new thread user creation.
 - `UserBase.statuses` now defaults to `None`. This will save a lot of memory when presence caching is enabled.
+- Add `UserBase.iter_guilds`.
 
 ##### ext.slash
 - Add `silent` parameter to `abort`.